requirements.txt
python-dotenv
requests
loguru
nanoid
aiosqlite
pytelegrambotapi
aiohttp
